requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
streamlit>=1.32.0
httpx>=0.25.0
plotly>=5.18.0